* Reverted rev 390 as it was causing crash on shutdown * Mutilate status changed to in front (by Grevs)
git-svn-id: svn://svn.aspiredev.org/AspireCore/trunk@442 a6a5f009-272a-4b40-a74d-5f9816a51f88

+++ src/ascent-shared/ByteBuffer.h
@@ -377,7 +377,7 @@ public:
 	void hexlike()
 		{
 			uint32 j = 1, k = 1;
-			printf("STORAGE_SIZE: %u\n", (uint32)size() );
+			printf("STORAGE_SIZE: %u\n", size() );
 			for(uint32 i = 0; i < size(); i++)
 			{
 				if ((i == (j*8)) && ((i != (k*16))))
